Fixed: https://github.com/vlachoudis/bCNC/pull/1714#issuecomment-1198337274 https://github.com/bosd/bCNC/pull/1#discussion_r932132554 https://github.com/bosd/bCNC/pull/1#discussion_r932140347
Also changed application to be inherited from tkinter.Tk, instead of
tkinter.Toplevel. This change also fixes the issue, that the windowname
is also displayed in the dock.

Further I moved bmain.main functionality back to __main__.main.
